+++ StereoKitC/material.cpp
@@ -97,7 +97,7 @@ void material_set_active(material_t material) {
 	shader_set_active    (material->shader);
 	shaderargs_set_data  (material->shader->args, material->args.buffer);
 	shaderargs_set_active(material->shader->args);
-	for (size_t i = 0; i < material->shader->tex_slots.tex_count; i++) {
+	for (int i = 0; i < material->shader->tex_slots.tex_count; i++) {
 		tex2d_set_active(material->args.textures[i], i);
 	}
 }
